refactor: menu에서 product 직접적인 참조 제거 및 menuValidator 생성
