+++ src/main/java/kitchenpos/application/MenuService.java
@@ -5,11 +5,10 @@ import kitchenpos.application.dto.MenuRequest;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.MenuValidator;
 import kitchenpos.domain.repository.MenuGroupRepository;
 import kitchenpos.domain.repository.MenuProductRepository;
 import kitchenpos.domain.repository.MenuRepository;
-import kitchenpos.domain.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,14 +20,14 @@ public class MenuService {
     private final MenuRepository menuRepository;
     private final MenuGroupRepository menuGroupRepository;
     private final MenuProductRepository menuProductRepository;
-    private final ProductRepository productRepository;
+    private final MenuValidator menuValidator;
 
     public MenuService(MenuRepository menuRepository, MenuGroupRepository menuGroupRepository,
-                       MenuProductRepository menuProductRepository, ProductRepository productRepository) {
+                       MenuProductRepository menuProductRepository, MenuValidator menuValidator) {
         this.menuRepository = menuRepository;
         this.menuGroupRepository = menuGroupRepository;
         this.menuProductRepository = menuProductRepository;
-        this.productRepository = productRepository;
+        this.menuValidator = menuValidator;
     }
 
     @Transactional
@@ -39,14 +38,13 @@ public class MenuService {
 
         final MenuGroup menuGroup = menuGroupRepository.findById(menuRequest.getMenuGroupId())
                 .orElseThrow(IllegalArgumentException::new);
-        final List<Product> products = mapToProducts(menuRequest);
-        final Menu menu = new Menu(menuRequest.getName(), menuRequest.getPrice(), menuGroup,
-                menuRequest.getMenuProducts(), products);
+        menuValidator.validatePriceByProducts(menuRequest.getPrice(), menuRequest.getMenuProducts(), menuRequest);
+
+        final Menu menu = new Menu(menuRequest.getName(), menuRequest.getPrice(), menuGroup, menuRequest.getMenuProducts());
         final Menu savedMenu = menuRepository.save(menu);
 
         final List<MenuProduct> savedMenuProducts = createMenuProducts(menuRequest, savedMenu);
-        return new Menu(savedMenu.getId(), savedMenu.getName(), savedMenu.getPrice(), savedMenu.getMenuGroup(),
-                savedMenuProducts, products);
+        return new Menu(savedMenu.getId(), savedMenu.getName(), savedMenu.getPrice(), savedMenu.getMenuGroup(), savedMenuProducts);
     }
 
     private List<MenuProduct> createMenuProducts(MenuRequest menuRequest, Menu savedMenu) {
@@ -59,18 +57,6 @@ public class MenuService {
 
     private Menu findMenu(Menu menu) {
         return menuRepository.findById(menu.getId())
-                .orElseThrow(IllegalArgumentException::new);
-    }
-
-    private List<Product> mapToProducts(MenuRequest menuRequest) {
-        return menuRequest.getMenuProducts()
-                .stream()
-                .map(this::findProductByMenuProduct)
-                .collect(Collectors.toList());
-    }
-
-    private Product findProductByMenuProduct(MenuProduct menuProduct) {
-        return productRepository.findById(menuProduct.getProductId())
                 .orElseThrow(IllegalArgumentException::new);
     }
