+++ src/main/java/kitchenpos/domain/Menu.java
@@ -3,10 +3,7 @@ package kitchenpos.domain;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.math.BigDecimal;
 import java.util.List;
-import java.util.stream.Collectors;
-import javax.persistence.CascadeType;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -37,14 +34,12 @@ public class Menu {
     }
 
     public Menu(final String name, final BigDecimal price, final MenuGroup menuGroup,
-                final List<MenuProduct> menuProducts, final List<Product> products) {
-        this(null, name, price, menuGroup, menuProducts, products);
+                final List<MenuProduct> menuProducts) {
+        this(null, name, price, menuGroup, menuProducts);
     }
 
-    public Menu(Long id, String name, BigDecimal price, MenuGroup menuGroup, List<MenuProduct> menuProducts,
-                List<Product> products) {
+    public Menu(Long id, String name, BigDecimal price, MenuGroup menuGroup, List<MenuProduct> menuProducts) {
         validatePrice(price);
-        validatePriceWithProducts(price, menuProducts, products);
         this.id = id;
         this.name = name;
         this.price = price;
@@ -60,44 +55,6 @@ public class Menu {
         if (price.compareTo(BigDecimal.ZERO) < 0) {
             throw new IllegalArgumentException("메뉴 가격이 0보다 작으면 예외가 발생한다.");
         }
-    }
-
-    private void validatePriceWithProducts(final BigDecimal price, List<MenuProduct> menuProducts,
-                                           List<Product> products) {
-
-        final List<Long> productIds = mapToProductIds(menuProducts);
-
-        BigDecimal sum = BigDecimal.ZERO;
-        for (final MenuProduct menuProduct : menuProducts) {
-            final Long productId = getProductIdByMenuProduct(productIds, menuProduct);
-
-            final Product product = getProductById(products, productId);
-            sum = sum.add(product.getPrice().multiply(BigDecimal.valueOf(menuProduct.getQuantity())));
-        }
-
-        if (price.compareTo(sum) > 0) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private Long getProductIdByMenuProduct(List<Long> productIds, MenuProduct menuProduct) {
-        return productIds.stream()
-                .filter(id -> id.equals(menuProduct.getProductId()))
-                .findAny()
-                .orElseThrow(() -> new IllegalArgumentException("product를 찾을 수 없습니다."));
-    }
-
-    private Product getProductById(List<Product> products, Long productId) {
-        return products.stream()
-                .filter(it -> it.getId().equals(productId))
-                .findAny()
-                .orElseThrow(IllegalArgumentException::new);
-    }
-
-    private List<Long> mapToProductIds(List<MenuProduct> menuProducts) {
-        return menuProducts.stream()
-                .map(MenuProduct::getProductId)
-                .collect(Collectors.toList());
     }
 
     public Long getId() {
