+++ src/main/java/kitchenpos/domain/MenuValidator.java
@@ -1,0 +1,10 @@
+package kitchenpos.domain;
+
+import java.math.BigDecimal;
+import java.util.List;
+import kitchenpos.application.dto.MenuRequest;
+
+public interface MenuValidator {
+
+    void validatePriceByProducts(BigDecimal price, List<MenuProduct> menuProducts, MenuRequest menuRequest);
+}
